Don't lose state for off-schedule displays
This change moves the filtering of unmonitored displays earlier in the
process so that displays that are still being monitored, but are
currently off-schedule will not be filtered. The presence check is only
done on the displays that are currently on-schedule. Those displays will
have their statuses updated and emails sent out as required.

See [issue
791](https://trello.com/c/aVAVyhQA/5104-2-defect-791-display-monitoring-not-sending-out-notifications-correctly)